v9.1: Mostly fixed bugs, like header to stick and mobile responsivness. Refactored StoryCard to two files.
