+++ styles/globals.css
@@ -23,8 +23,6 @@ html {
 body {
   line-height: 1.5;
   -webkit-font-smoothing: antialiased;
-  overflow-x: hidden;
-  /* overflow-y: scroll; */
 }
 
 img,
